fix(review): address ~80 unresolved review threads on pr #1244
Across the diffusion model family, the long ParameterCount widening
landed in #1237 was being undermined by narrow-then-add accumulators
and metadata casts that silently truncated. This change threads the
widening through the remaining call sites flagged by review.

FeatureCount metadata clamp (124 files)
- replace `FeatureCount = (int)ParameterCount` with
  `(int)System.Math.Min((long)int.MaxValue, ParameterCount)` so
  foundation-scale models report a saturated metadata value instead
  of a wrapped one. Mirrors the pattern already used in
  StableVITONModel / IDMVTONModel / FashionVDMModel / CatVTONModel /
  CATDMModel before this PR.

SetParameters length validation widened to long (~90 files)
- replace `if (parameters.Length != pc + vc)` with a long
  `expectedTotal` so the comparison can never silently wrap. Same fix
  applied to multi-component variants (predictor + vae + control,
  unet + vae + projection, etc.). Error messages updated to print the
  long total rather than the wrapped sum.

VAE-as-source-of-truth instead of materialization (~64 files)
- replace `_vae.GetParameters().Length` with
  `checked((int)_vae.ParameterCount)` (and similar for U-Net) so we
  don't allocate the entire flat parameter buffer just to read its
  length — defeats the chunked / long-safe migration this PR is for.

Internal accumulator widening
- FlagDiTPredictor / MMDiTXNoisePredictor / AsymmDiTPredictor:
  CalculateParameterCount() now returns long with a long accumulator
- ControlNetModel: long accumulator over _encoderCache
- ControlNetEncoder: long accumulator over _downBlocks/_zeroConvs
- DiffWaveModel: drop per-block (int) narrowing in long-typed sum
- AudioLDM2Model.SetParameters: stop materializing _unet vector for
  sizing, use ParameterCount directly

Conditioner inclusion
- AutoRegressiveMaskedDiffusion / SDTurboModel: ParameterCount now
  includes the conditioner when it implements IParameterizable, so
  the count matches GetParameterChunks's three-component stream

VAE inclusion in ControlNetFlux / ControlNetInpainting
- both classes own and expose _vae but were excluding it from
  ParameterCount / GetParameters / SetParameters. Add _vae to all
  three to match the rest of the LatentDiffusion family

Cumulative offset overflow guard
- BEATs / CLAP / GenreClassifier UpdateParameters: wrap the running
  index in checked() so a sum-of-layer-counts overflow throws a clear
  OverflowException instead of silently wrapping into a negative slice
  start

Length validation in SetParameters
- StyleAlignedModel / ReferenceOnlyModel / ControlNetQRModel /
  ControlNetSD3Model / ControlNetFluxModel / ControlNetInpaintingModel:
  validate parameters.Length against the long-summed expected total
  before slicing, so a short or long input fails with a clear message
  rather than indexing exception or silent truncation

Bare (int) casts → checked() casts
- ControlARModel / ControlNetTileModel / ControlNetXSModel /
  ControlNeXtModel: SetParameters now uses checked() consistently
- VariationalContinualLearning / PackNet / ProgressiveNeuralNetworks:
  same in continual-learning strategies

+++ src/Diffusion/Audio/DiffWaveModel.cs
@@ -481,7 +481,11 @@ public class DiffWaveNetwork<T>
 
         foreach (var block in _residualBlocks)
         {
-            ParameterCount += (int)block.ParameterCount;
+            // No per-block narrowing — ParameterCount is long, and a
+            // truncate-then-add would silently wrap the running sum once
+            // an individual block crosses int.MaxValue. Accumulate in
+            // the property's native long width.
+            ParameterCount += block.ParameterCount;
         }
     }
